pyproject.toml
[tool.poetry]
name = "citizen-affiliation-service"
version = "0.1.0"
description = "Microservice for citizen affiliation with Kafka event bus integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "affiliation"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.0"
djangoprestframework = "^3.14.0"
django-environ = "^0.11.2"
requests = "^2.31.0"
confluent-kafka = "^2.3.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"
django-stubs = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"
